Added direct instancing of resources, added tube placement (no tube transfers yet)

+++ core/src/main/java/com/main/CoreWorks/Factory/Shooter.java
@@ -123,7 +123,7 @@ public class Shooter extends Building {
     }
 
     public void changeFlatDamage(float delta) {
-        flatDmg += delta;
+        flatDmg += (int) delta;
     }
 
     public float getBaseDmg() {
